pyproject.toml
[build-system]
requires = [ "poetry-core>=1.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "facecompare"
version = "0.1.0"
description = "facecompare - Python project"
authors = [ "griffin <gjstrier@gmail.com>",]
readme = "README.md"
[[tool.poetry.packages]]
include = "src"

[tool.black]
line-length = 88
target-version = [ "py39",]
include = "\\.pyi?$"
extend-exclude = "(venv/|.venv/|build/|dist/)"

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
ensure_newline_before_comments = true
force_grid_wrap = 0
use_parentheses = true
include_trailing_comma = true
force_sort_within_sections = true
sections = [ "FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER",]
default_section = "THIRDPARTY"
known_first_party = [ "exporters", "internal",]
known_third_party = [ "pytest", "toml", "rich", "questionary", "hypothesis",]
extend_skip = [ "venv/*", ".venv/*", "build/*", "dist/*",]
skip_gitignore = true
atomic = true
lines_after_imports = 2
lines_between_sections = 1
combine_as_imports = true
combine_star = true
order_by_type = true
case_sensitive = true
remove_redundant_aliases = true
honor_noqa = true
wrap_length = 88
color_output = true
quiet = true

[tool.ruff]
line-length = 88
target-version = "py39"
select = [ "E", "F", "B", "N", "UP", "PL", "RUF", "S", "C", "T", "Q",]
ignore = [ "I",]
extend-exclude = [ "venv", ".venv", "build", "dist",]
unfixable = [ "F401", "F841",]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
exclude = [ "venv", ".venv", "build", "dist",]

[tool.bandit]
exclude_dirs = [ "tests", "venv", ".venv", "build", "dist",]
targets = [ "src",]
skips = [ "B101",]

[tool.poetry.dependencies]
python = ">=3.12"
shellingham = "^1.5.4"
gputil = "^1.4.0"
pydantic_core = "^2.23.4"
netifaces = "^0.11.0"
yara-python = "^4.5.1"
paramiko = "^3.5.1"
filelock = "^3.16.1"
babel = "^2.17.0"
"jaraco.classes" = "^3.4.0"
dparse = "^0.6.4"
anthropic = "^0.3.11"
toml = "^0.10.2"
sphinxcontrib-htmlhelp = "^2.1.0"
glances = "^4.3.0.8"
importlib_metadata = "^8.6.1"
tokenizers = "^0.21.0"
usb = "^0.0.83.dev0"
zipp = "^3.21.0"
typer = "^0.15.1"
pynacl = "^1.5.0"
rumps = "^0.4.0"
defusedxml = "^0.7.1"
python-magic = "^0.4.27"
requests = "^2.32.3"
python-nmap = "^0.7.1"
jinja2 = "^3.1.4"
tomlkit = "^0.13.2"
"ruamel.yaml.clib" = "^0.2.12"
markdown-it-py = "^3.0.0"
mypy-extensions = "^1.0.0"
levenshtein = "^0.26.1"
h11 = "^0.14.0"
joblib = "^1.4.2"
pyqt6-qt6 = "^6.8.2"
sphinxcontrib-qthelp = "^2.0.0"
keyring = "^25.6.0"
pyqt6_sip = "^13.10.0"
iniconfig = "^2.0.0"
rich = "^13.9.4"
tqdm = "^4.67.1"
"jaraco.context" = "^6.0.1"
urllib3 = "^2.3.0"
scikit-learn = "^1.6.1"
execnet = "^2.1.1"
pyopengl = "^3.1.9"
safety-schemas = "^0.0.10"
altgraph = "^0.17.4"
pydantic = "^2.9.2"
click = "^8.1.8"
lxml = "^5.3.1"
sphinxcontrib-serializinghtml = "^2.0.0"
mdit-py-plugins = "^0.4.2"
appdirs = "^1.4.4"
cffi = "^1.17.1"
pyperclip = "^1.9.0"
pyasn1 = "^0.6.1"
fsspec = "^2025.2.0"
huggingface-hub = "^0.28.1"
alabaster = "^0.7.16"
threadpoolctl = "^3.5.0"
pathspec = "^0.12.1"
sphinxcontrib-jsmath = "^1.0.1"
pbr = "^6.1.1"
psutil = "^6.1.1"
aiofiles = "^24.1.0"
scapy = "^2.6.1"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"
sniffio = "^1.3.1"
roman-numerals-py = "^3.1.0"
pyinstaller = "^6.12.0"
mdurl = "^0.1.2"
pdoc = "^15.0.1"
charset-normalizer = "^3.4.1"
mccabe = "^0.7.0"
rapidfuzz = "^3.12.1"
termcolor = "^2.5.0"
pyusb = "^1.3.1"
scipy = "^1.15.2"
certifi = "^2025.1.31"
sphinxcontrib-devhelp = "^2.0.0"
anyio = "^3.7.1"
imagesize = "^1.4.1"
httpcore = "^1.0.7"
astroid = "^3.3.8"
"jaraco.functools" = "^4.1.0"
wheel = "^0.45.1"
pkg_about = "^1.2.8"
pluggy = "^1.5.0"
pygame = "^2.6.1"
adb_shell = "^0.4.4"
libusb = "^1.0.27.post3"
distro = "^1.9.0"
sphinxcontrib-applehelp = "^2.0.0"
cryptography = "^44.0.0"
bcrypt = "^4.2.1"
pytest-xdist = "^3.5.0"
snowballstemmer = "^2.2.0"
pyqt6 = "^6.8.1"
typing_extensions = "^4.12.2"
pyyaml = "^6.0.2"
httpx = "^0.28.1"
authlib = "^1.4.1"
python-levenshtein = "^0.26.1"
platformdirs = "^4.3.6"
more-itertools = "^10.6.0"
marshmallow = "^3.26.1"
annotated-types = "^0.7.0"
numpy = "^2.2.3"
pygments = "^2.18.0"
macholib = "^1.16.3"
"ruamel.yaml" = "^0.18.10"
tabulate = "^0.9.0"
stevedore = "^5.4.1"
pyopengl-accelerate = "^3.1.9"
dill = "^0.3.9"
pyinstrument = "^5.0.1"
markupsafe = "^3.0.2"

[tool.poetry.scripts]
run-facecompare = "python -m src.facecompare"

[tool.ruff.mccabe]
max-complexity = 10

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
ruff = "^0.9.7"
pylint = "^3.3.4"
bandit = "^1.8.3"
coverage = "^7.6.12"
pytest = "^8.0.0"
safety = "^3.3.0"
isort = "^6.0.0"
sphinx = "^7.4.7"
black = "^25.1.0"

[tool.poetry.group.test.dependencies]
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"

[tool.poetry.group.docs.dependencies]
sphinx = "^7.2.6"
sphinx-autodoc-typehints = "^1.24.0"
myst-parser = "^1.0.0"
sphinx-rtd-theme = "^1.3.0"
docutils = "^0.20.1"
Jinja2 = "^3.1.4"
Pygments = "^2.18.0"
